Remove companion objects from H_OpaqueTypesWithCompileError.scala to reduce distractions; Improve error messages; Include scaladoc
